refactor(agent-core): split task-intelligence into source-resolver and budget-calculator (#1647)
